requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
langchain==0.0.311
langchain-openai==0.0.2
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.4.2
piper-tts==1.2.0
pydub==0.25.1
ffmpeg-python==0.2.0
pytest==7.4.2
httpx==0.24.1
python-multipart==0.0.6
jinja2==3.1.2 